[FIRRTL] Replace Lower Types with new version.
